@@ -25,30 +25,29 @@
 using Android.Content;
 using Android.Graphics;
 
-namespace Sharpnado.MaterialFrame.Droid
+namespace Sharpnado.MaterialFrame.Droid;
+
+public interface IBlurImpl
 {
-    public interface IBlurImpl
+    bool Prepare(Context context, Bitmap buffer, float radius);
+
+    void Release();
+
+    void Blur(Bitmap input, Bitmap output);
+}
+
+public class EmptyBlurImpl : IBlurImpl
+{
+    public bool Prepare(Context context, Bitmap buffer, float radius)
     {
-        bool Prepare(Context context, Bitmap buffer, float radius);
-
-        void Release();
-
-        void Blur(Bitmap input, Bitmap output);
+        return false;
     }
 
-    public class EmptyBlurImpl : IBlurImpl
+    public void Release()
     {
-        public bool Prepare(Context context, Bitmap buffer, float radius)
-        {
-            return false;
-        }
+    }
 
-        public void Release()
-        {
-        }
-
-        public void Blur(Bitmap input, Bitmap output)
-        {
-        }
+    public void Blur(Bitmap input, Bitmap output)
+    {
     }
 }
