@@ -4,7 +4,7 @@ namespace Sharpnado.MaterialFrame.Droid
 {
     internal static class JniExtensions
     {
-        public static bool IsNullOrDisposed(this Java.Lang.Object javaObject)
+        public static bool IsNullOrDisposed(this Java.Lang.Object? javaObject)
         {
             return javaObject == null || javaObject.Handle == IntPtr.Zero;
         }
